Fix for spoqify delay
The first load of a spoqify playlist seems to always be empty. A refresh fixes it in the browser but no way to do that in the client with Spicetify. So send the client to a dummy url and then send them to the proper one after a short delay.